context: use size_t for lengths

+++ include/nmea/context.h
@@ -18,6 +18,8 @@
 #ifndef __NMEA_CONTEXT_H__
 #define __NMEA_CONTEXT_H__
 
+#include <stddef.h>
+
 #ifdef  __cplusplus
 extern "C" {
 #endif /* __cplusplus */
@@ -34,7 +36,7 @@ extern "C" {
  * @param s the string to trace
  * @param sz the length of the string
  */
-typedef void (*nmeaTraceFunc)(const char *s, int sz);
+typedef void (*nmeaTraceFunc)(const char *s, size_t sz);
 
 /**
  * Function type definition for error logging
@@ -42,15 +44,15 @@ typedef void (*nmeaTraceFunc)(const char *s, int sz);
  * @param s the string to log
  * @param sz the length of the string
  */
-typedef void (*nmeaErrorFunc)(const char *s, int sz);
+typedef void (*nmeaErrorFunc)(const char *s, size_t sz);
 
 void nmea_context_set_trace_func(nmeaTraceFunc func);
 void nmea_context_set_error_func(nmeaErrorFunc func);
-void nmea_context_set_buffer_size(int sz);
-int nmea_context_get_buffer_size(void);
+void nmea_context_set_buffer_size(size_t sz);
+size_t nmea_context_get_buffer_size(void);
 
 void nmea_trace(const char *s, ...) __attribute__ ((format(printf, 1, 2)));
-void nmea_trace_buff(const char *s, int sz);
+void nmea_trace_buff(const char *s, size_t sz);
 void nmea_error(const char *s, ...) __attribute__ ((format(printf, 1, 2)));
 
 #ifdef  __cplusplus
